detect/http_method: move all tests into tests/
